EES-1531 Fix cascading delete not working for Subject.Filters
This fixes a regression in the cascading delete of `Subjects` due to the
addition of the `Subject.Filters` property. Previously, we thought that
detaching the `Subject` from the context would be enough to allow the
correct deletion, however, this has turned out to be untrue.

With the new `Filters` property, EF attempts to delete the related
`Filters` first (this is a client cascading delete), but is unable to
do so because of existing `ObservationFilterItems` still related to
`Filter`. This throws a SQL foreign key constraint error.

Instead, we now just execute the delete as raw SQL to bypass EF's
client cascading delete behaviour.