EES-1727 Disable ModalConfirm buttons and exits when confirming/cancelling
This change prevents `ModalConfirm` buttons from being clicked and the
modal from being exited (e.g. with Esc or an underlay click) whilst the
current `onConfirm` or `onCancel` is in progress.

Whilst confirming/cancelling is in progress, the user should be unable
to do anything to prevent them from getting the page into an
inconsistent state.