EES-1754 Refactor section block components to avoid coupling with releases
This change refactors all the various section block components across
the project to remove their current dependence on release-specific
props. The crux of the issue was caused by dependencies on a common
`SectionBlocks` component that extended its props from `DataBlockTabs`.

`DataBlockTabs` itself is coupled with the concept of releases, meaning
that this coupling was propagated to all `SectionBlocks` consuming
components. We have now removed `SectionBlocks` completely and have
replaced this with context-specific implementations e.g.
`PublicationSectionBlocks` and `MethodologySecitonBlocks`.

In the admin, `EditableSectionBlocks` has been refactored to allow it to
more generically render different types of blocks. This provides enough
flexibility to correctly render the blocks for release or methodology
contexts without the tight coupling with releases.