EES-1767 Fix chart builder boundary level field not being initialized
This change fixes the chart builder's boundary level field (for maps)
not being initialized with a value (regardless of if you save the chart
with a different boundary level or not).

This was due to `boundaryLevel` not being injected into the
`chartBuilderReducer` (from the data block's table data query). Instead,
we have now completely removed the boundary level field from the
`ChartConfiguration` form and `chartBuilderReducer`. It is now a
standalone `FormSelect` controlled by the `boundaryLevel` prop.