EES-1804 Fix legend item data sets not saving with optional properties
This change adds a new `ChartLegendItemDataSet` type which allows all of
the constitutent properties to be optional. This is important as the
frontend will typically remove the grouping filter type from the data
set.

Previously this was causing empty guids to be persisted in the data sets
when grouping by indicator. These would then be returned in the response
and the data sets could not be matched correctly to the data.