EES-1838 Change data block list endpoint to return summary view models
This changes the `DataBlocksController.List` endpoint to return a list
of `DataBlockSummaryViewModels` instead of `DataBlockViewModels`. This
summary variant includes a `ChartsCount` and the `ContentSectionId`,
allowing us to display more information about the data blocks in the
list page.

This response is also smaller due to not including the bulkier fields
like `charts`, `table` and `query`.