EES-1881 Refactor out TableBuilderReleaseController for admin
This change extracts a new `TableBuilderReleaseController` out of
`TableBuilderMetaController`. This allows us to improve the REST
semantics and avoid naming collisions over what is and isn't 'meta'.

As part of this, we've moved  `ReleaseService` (renamed from
`ReleaseMetaService`) into the `Data.Services` project. This is a more
appropriate location for the service, as it allows us to re-use it in
other projects more easily.