EES-1713 Add custom CKEditor 5 build to project with ImageResize plugin
This change adds a new `explore-education-statistics-ckeditor`
subproject that contains a custom CKEditor 5 build. We now use this
instead of the default classic build.

We have gone with a custom build as it saves a bunch of time compared to
importing individual CKEditor packages into an ad-hoc editor instance.
Frustratingly, this is actually quite a lot of work without type
definitions.

Creating the custom build was also really easy thanks to CKEditor's
online builder: https://ckeditor.com/ckeditor-5/online-builder/