EES-289 Refactor release pages to make single getRelease request
This changes refactors the existing `ManageReleaseContext` to the
simpler `ReleaseContext` and embeds a release summary inside of this
when the `ReleasePageContainer` is rendered.

This reduces the number of requests we have to make as the release
summary provides all of the details we need for all the various pages
and their dependent components.