EES-2030 Add Created to ContentBlock
This commit adds Created to the Content DB ContentBlock table.

It also saves a created date whenever a new ContentBlock is created. But
it only returns the created date to the frontend when returning summary info about DataBlocks.

The reason for saving the info for all blocks but only returning for
DataBlockSummaryViewModel is that the field has to be added to ContentBlock table, so
there is no extra cost to saving the created date for all blocks, even
if we're only currently using it in one specific instance.

NOTE: It is saving the Created date in blocks created in the Content and
Annexes fields of the Methodologies table.