EES-1526 Create ReleaseStatus.cs for the ContentDB
This commit creates ReleaseStatus.cs, which will eventually replace the
ReleaseStatus Azure storage table. So while the work for EES-1526 is
to save older InternalReleaseNotes, we're doing this work in such a way
that it will make the future work of removing the ReleaseStatus Azure
storage table easier.

To make the migration of data from the ReleaseStatus Azure storage
table to the ContentDB easier, we're also passing the ReleaseStatusId
of any new entry created in the ContentDB to the publisher so the Azure
storage table has the same Id as its RowKey.