EES-1812 Fix drag event leaking to parent row/col group with keyboard
This fixes a weird dragging issue that occurs when:

- User starts 'dragging' row/col items within a row/col group using
  their keyboard i.e. they focus an item and then press Space.
- User re-orders the item using arrow keys, then finishes the 'drag'
  event by pressing Space again.
- The items are re-ordered, but a new 'drag' event starts immediately on
  the parent row/col group itself (not the child items).

This is pretty undesirable behaviour as it makes it really hard to use
the keyboard correctly. This is most likely impacting on the ability to
test these components in the UI tests.

It turns out this was a similar case to when click events would
sometimes start drag events on the row/col group instead of the row/col
item.

To fix this, we've just applied a similar hack to the click events. This
time we've used the `blur` and `focus` events instead to disable
dragging the  parent row/col group  whilst the focus is within the group
itself.