EES-1812 Refactor TableTool components tests to not use page snapshots
Whole page snapshots are a really bad idea as they will often pick up
changes unrelated to what the test case is trying to assert (such as
with these tests).

It's better to explicitly assert what we're looking for, or really
narrow down the scope of snapshot to a specific part of the page.