EES-2536 Fix custom selectors failing for negation keywords
This fixes cases where we are using keywords like `user checks page does
not contain`. These were primarily failing due to keywords that were
waiting for elements that would never appear.

It also turns out that we need to return lists of elements for custom
selectors instead of singular elements.