EES-1737 Move publication trees from publisher into Content API
This removes the caching of the publication content trees in the
publisher in favour of generating the trees in the content API upon
request.

As part of this, we've added a lot more tests around the generation
of the content trees.

There may be performance implications to this, which we'll be addressing
in a later commit.