EES-1944 Convert BlobContainers to records
We've opted to convert these to records as there are some
disadvantages to using structs that make them awkward for how we
want the upcoming caching functionality to work.

In particular - structs always have a default, paramaterless
constructor that is actually invisible to reflection (!!!).
This means they add an additional layer of complication when it
comes to figuring out which constructor to use for the cache key.