EES-2548 Re-work `ICacheService` into a generic caching interface
The previous `ICacheService` interface was dependent on specific details
about blob storage for its caching key. This meant it couldn't be used
for different types of caches in the future (memory, Redis, etc).

We've now re-worked this so that `ICacheService` is generic, allowing us
to use different `ICacheKey` subtypes such as an `IBlobCacheKey`.

`IBlobCacheKey` contains both the blob container name and the key,
and is used by `IBlobStorageService` for its blob storage specific
implementation.

This work is particularly important to allow us to build on-top of it in
the upcoming AOP-based caching implementation.