EES-2548 Add non-generic `ICacheService.GetItem` method
We have now added a variant of `ICacheService.GetItem` that does not
have a type parameter, but allows consumers to pass through an actual
`Type` value instead.

This is advantageous as it allows us to have an `object?` return type,
which will be useful for the upcoming AOP caching work.