EES-2548 Add AOP based caching attributes
This change adds AOP based caching using attributes (and a fair bit of
reflection). To enable this we have used the Aspect Injector library so
that we do all of the aspect injection work at compile time (rather than
at runtime with dynamic proxy based implementations).

We can now really easily add caching by annotating methods with
implementations of `CacheAttribute` e.g. `BlobCacheAttribute`. These
must be assigned a cache key (types derived from `ICacheKey`) so that
we can try and construct it and pass it to the relevant caching service
e.g. `BlobCacheService`.

It should be noted that we perform a fair amount of reflection to get
figure out how to construct the cache key using the invoked method's
parameters. There may be additional performance considerations for
this in the future.