EES-2065 To publish amendments requires new release note
This commit fixes a bug where an amendment could be published, even if
it didn't have a new release note. Amendments were publishable if they
had a release note from a previous version.

To fix this bug we've needed to add a Created date to the Update table,
which is compared to the Created date of the current release to ensure a
new release note exists.

We've also taken this opportunity to add a CreatedBy field to the Update
table.