EES-2236 Fix missing zip file permissions in Unix environments
Currently, .NET doesn't automatically add Unix file permissions to our
zip's file entries. This means that all of the files cannot be opened
without setting the permissions manually in the OS.

To fix this, we manually add the Unix file permissions ourselves by
converting a permission string to an octal literal on the zip entry.

I believe that in .NET 6, this should be done automatically for us due
to the following change: https://github.com/dotnet/runtime/pull/55531