EES-2581 Record prerelease invite even if user is already on the system
Previously, if a preexisting user on the system was invited to be a
prerelease user for a release, we just added an entry in
UserReleaseRoles for them and didn't create a new row in
UserInvites/UserReleaseInvites. However, if a user didn't exist on
the system, their associated UserInvites/UserReleaseInvites would
remain on the system, even after their user is created.

This change means that we now always create a record of a prerelease
invite in UserInvites/UserReleaseInvites, even if that permission is applied or
"Accepted" instantly.

While this may be useful for auditing purposes, the primary motivation
for this change is to allow us to record if a prerelease invite email has
been sent in one location. Now that all prerelease users for a release will have an
associated UserReleaseInvites row, we can now record whether the email has been
sent in that table.