EES-2658 Split out subjects and featured tables into separate endpoints
This refactors the Data API's previous endpoint for retrieving a
release/publication's subjects and featured tables (previously table
highlights) into two separate endpoints.

We are doing this in preparation for use in the new data catalogue page,
as we want to be able to retrieve the subjects independently for the
final step of the wizard.