EES-2658 Add file details to `Data.Services.SubjectViewModel`
This change adds `FileInfo` details to `SubjectViewModel`. We have
currently implemented this via a temporary `IBlobInfoGetter`
interface.

We hope to be able to remove this with EES-2343, when we can store the
file size inside the database itself (rather than in blob storage).
This should allow us to always be able to construct the `FileInfo`,
without needing to fetch a corresponding `BlobInfo`.