EES-2369 Remove frontend ordering of releases in data catalogue
This change removes the incorrect frontend ordering of releases as
ordering should not be dependent on the published date (this can be
completely different to the time period a release relates to).

Instead, we just the order provided by the backend as this should be
used as the source of truth.