EES-2007 Replace data files with data catalogue and all files zip links
This reworks the current 'Explore data and files' accordion section so
that we remove the list of data files, unless there is no data
catalogue to link to (e.g. in the admin).

As the data files list has been removed, we now need to make the all
files zip link more prominent. We have done this by making it a right
aligned button similar to the 'Create tables' button.