EES-2546 Replace 'All files' zip endpoint with dynamically generated zip
This change replaces the existing all files zip (`/files/all`) endpoint
in favour of dynamically generating zip files using the new `/files`
endpoint.

To generate this file using the `/files` endpoint, we simply don't
provide any `fileIds` to filter on i.e. all files should be returned.

As part of this, we cache the dynamically all files zip for up to an
hour to hopefully avoid any potential performance issues from
generating this file every time.

It should be noted that no caching is done if `fileIds` are provided.
This would require caching of each combination of `fileIds` and it
didn't seem worthwhile to do so.