EES-2846 Fix UI tests
This fix is a workaround for a CKEditor bug. Usually, after adding an
image, there are buttons on the image outline to add a paragraph before
or after the image. The bug is that if the text cursor is immediately
before some text and you *then* add an image, these buttons to add a
paragraph above/below the image don't appear.

It is still possible to add lines above and below and image, so it was
decided that we'd change the UI tests to work around this bug rather
than downgrade CKEditor.