EES-2128 Fix broken data catalogue downloads due to undefined content API URL
When using `downloadService.downloadFiles`, it appears that
`CONTENT_API_BASE_URL` is undefined at the time that we try and download
the data catalogue files.

It seems like changes in Next.js v11 (or `next-transpile-modules`) means
that modifications to `process.env` made by the frontend's `loadEnv`
are not being respected.

Consequently, we've gone with a simple fix (for now) where we just use
the `baseURL` used by the content API (seems to get set by setting the
Axios defaults in `_app.ts`).

In the future, this could be a potential source of bugs, so we might
want to do some further work to pull environment variables into common
without using `process.env`.