EES-2128 Remove `as` prop from frontend `Link` component
Since Next v10, the `next/link` component now just auto-resolves `href`,
instead of needing to pass it both a `href` and an `as` prop.

Consequently, we can just completely remove the `as` prop as it is
redundant and even seems to be harmful in some cases (i.e. causing
Webpack errors when navigating between pages in development).