EES-2893 Add basic auth credentials into request headers in UI tests
This changes how the UI tests perform basic authentication against the
public frontend. Previously, we were putting these into the `PUBLIC_URL`
variable, but unfortunately this seems to now be incompatible with
Next.js now that we've upgrade to v11.

To get around this, we now add the basic auth credentials into the
request headers using the Chrome DevTools prototocol.