EES-2773 Fix `LocationAttributeNode` having nullable `Attribute`
This addresses null reference warnings due to `LocationAttributeNode`
having an incorrect nullable `Attribute` property. It is not actually
possible for this to be null as we extract the various
`LocationAttributes` from non-null getters in `Location`.

It also doesn't make sense for `Attribute` to be null in the context of
what `LocationAttributeNode` is supposed to do!

We have addressed this by refactoring `LocationAttributeNode` to require
an attribute upon initialization using its constructor.