EES-2922 Add `passOnce` prop to `Gate` component
This adds a new `passOnce` prop to `Gate` to allow us to explicitly
control if the gate should 'close' as soon as the condition passes, or
it should allow the condition to toggle between true/false.

We have also refactored `Gate` so that it's more obvious what is going
on. We were previously using `useAsyncCallback`, but the code required
to make this work was quite complex.

Consequently, we've refactored `Gate` to use a simple `useEffect` that
handles our async state in a more precise way, making the code more
understandable.