EES-2777 Update table data meta to handle hierarchical locations
This change updates `TableDataSubjectMeta` to swap its `locations` key
with `locationsHierarchical` and `mapFullTableMeta` to parse this
correctly.

As part of this, we've needed to re-work some of the location
deduplication code to get it working with the new data structure
presented by `locationsHierarchical`. It was a bit burdensome to have to
make these changes, but We should hopefully be able to remove this
code entirely pending the upcoming location IDs work.