EES-2912 Fix unmounted state change errors when `Button` is clicked
This was being caused by recent update to `Button` that disabled it
when an async task was being performed in its `onClick`. This could
potentially lead to `toggleClicking.off` being called after the
component had already unmounted as this component is typically used
in contexts where the page might change after the button is clicked.