EES-3012 Update `BlobStorageService.DeleteBlobs` with include regex option
This change refactors the `DeleteBlobs` method so that it accepts a new
options object that has inclusion/exclusion regexes that can be used to
filter the blobs that are deleted.

To avoid changing the current API significantly (and having to refactor
all the call sites), we've also avoided adding the `directoryPath` param
into the options class for the timebeing.