EES-2823 Fix `EditableContentBlock` always autosaving (even when disabled)
This was noticed in methodologies, where `MethodologyEditableBlock`
would trigger a save despite having the `autoSave` prop being set to
false.

To fix this, we've reworked the autosaving functionality to work around
a new `onAutoSave` handler on `EditableContentForm`. This differentiates
it from the `onSubmit` handler, which is typically only called when
a button is pressed. Previously, both auto-saving and submitting were
passed through the same handler, leading to this bug.