EES-2823 Refactor content context calls into `ReleaseEditableBlock`
This change removes the `onSave` and `onDelete` props from
`ReleaseEditableBlock`, allowing us to move the respective
`ReleaseContentContext` calls into `ReleaseEditableBlock` itself.

This reduces the amount of duplicate code and sets the precedent to
interact with the context in `ReleaseEditableBlock`.

As part of this, we've also cleaned up some of the naming in
`ReleaseContentContext` to make it consistent.