EES-2823 Add `Tooltip` component using Popper.js
We have implemented this without `react-popper` as there were some
noticable performance issues with it when using virtual elements
(i.e. with `followMouse = true`). This seemed to be due to updates
on `mouseover` events causing too may re-renders for React to handle.

Instead, we manage our own Popper instance and allow it to directly
manipulate the DOM without re-rendering the `Tooltip` component in
React. This is far more performant and leads to a much nicer
experience.