EES-2823 Remove element selector from `user presses keys` keyword
This change removes the selector argument from `user presses keys` as
this keyword was never intended to target an element.

We have now replaced any keywords using this old API with the more
suitable `user enters text into element` keyword, and changed the keys
argument to be variable (to allow combinations of keys).