EES-3019 Remove full table title button from `DataTableCaption`
This change removes the 'View/hide full table title' button that was
previously in `DataTableCaption`. We've opted to remove this as it's
unrealistic that a user would ever find the full table title useful once
it exceeds the truncation limits.

The previous behaviour was also poor for accessibility as screen readers
could end up reading very long titles, which can be very repetitive and
isn't very useful.

We now just show the generated 'short' title as the default.