EES-3281 Make chart reference lines more visible
This change improves the visibility of chart reference lines by making
them wider, darker and dashed.

As part of this, we have also refactored the charts to use a
`createReferenceLine` util function. Note that we have not used a
component as Recharts does not render non-Recharts child components
within a chart context.