EES-3281 Fix chart builder form submits using stale state
Form submits can currently occur with stale state due to the chart
builder state not being updated in time for when the actual submit
occurs. To mitigate this, we've used a debounced callback with a 300ms
delay that should hopefully mean that the chart builder state updates
in time for most cases.

Ultimately, we want to think about rearchitecting chart builder's state
management to ensure it is synchronized correctly - see EES-3355.