EES-3479 Fix reference line position showing empty options
This changes the 'Position' field in `ChartReferenceLinesConfiguration`
so that it no longer displays every single option from table meta. This
was particularly unwieldy in cases where there were options that did not
actually appear in the chart.

Instead, we have now added a `dataSetCategories` prop to determine which
options will be available. These are passed in from the parent
`ChartAxisConfiguration` component.