EES-3554 Tweak pagination view models and default query params
This twewaks the existing `PaginatedListViewModel` construction so that
pagination can be performed at the database as well as in consumers.

The previous API for this only really allowed consumers to paginate on
all the results that had already been pulled into memory. In most cases,
we want to try and avoid this, with the new `Paginate` static method
only being used as a fallback when its not possible to paginate in the
database.

For publication releases, we would need to re-model `Releases` to allow
us to paginate in the database (see EES-3663).