EES-3554 Update release create/update requests to use int `Year`
This refactors the release create/update request types so that the
release's time period year is set via a `Year` field, instead of the
weird `ReleaseName` field.

A further extension of this would be to also refactor the
`Release.ReleaseName` column in the database to `Release.Year` (as
an integer type instead of a string).