EES-3554 Integrate `listReleases` endpoint with PublicationReleasesPage
This integration is implemented using React Query. We have removed the
WIP that was previously using the old `publicationService.getMyPublication`
endpoint to fetch the publicaiton's releases.

As part of this, we have had to restructure most of the existing code
to allow us to run two queries upon page load for published and
unpublished releases. This has mostly manifested as a new
`PublicationUnpublishedReleases` component to handled unpublished
releases.