EES-3607 Various tidy up across new Find Statistics page
This mostly cleans up various parts of the data loading and rendered
page (e.g. removing an unnecessary wrapping fragment).

We have also added a new `isOneOf` type guard that's useful for
asserting that a value is one of a set of items. Great for checking that
a string is a string literal.