EES-3779 - amended way that Filters, FilterItems, FilterGroups and IndicatorGroups get constructed, to only require their owning entity id rather than the entity itself.
This helps in stopping EF in re-saving these linked entities when using new DbContext delegate pattern (which has no notion that, for instance, a FilterGroups owning Filter is already saved in the database).  Previously this was fine as we were saving a FilterGroup in the same DbContext that its owning Filter was retrieved from, but using the new delegate pattern, these are 2 separate DbContexts.