EES-4192 - removed ImporterFilterCache and ImporterFilterService as we no longer need a distributed cache for sharing amongst different parallel batch importer Function executions.
Added DataFileReader and MetaDataFileReader classes for caching of CSV column indexes and more efficient reading of CSV rows.

Added up-front retrieval and caching of Filters, FilterGroups and FilterItems at the start of Stage 3 (Import Observations) instead of retrieving them individually as and when they are encountered while scanning the file, saving considerable time for files with large sets of FilterItems.