EES-4444 Refactor `Client` to hide Axios implementation details
This refactors `Client` so that the underlying Axios instance is
encapsulated more as an implementation detail. Of course, this isn't a
complete solution and is a bit leaky in some cases, but should be good
enough for the time being.

Importantly, we can now provide a `rawResponse: true` config option that
provides the raw underlying response object (instead of having to do
this through the Axios instance).