EES-4577 Refactor `RHFFormField` to only use `as` prop
This refactors `RHFFormField` to only use the `as` prop for component
rendering and reworks the typing so this is stricter.

Previously, the typing was too permissive and did not prevent the use of
invalid component types. We have now rectified it, and we are more
likely to get compiler errors when a non-form component is used.