EES-4599 Make spacing between blocks more consistent in HTML to text
This improves the current spacing implementation so that spacings are
more consistently added between blocks and their siblings. This aims to
ensure we only add 2 line breaks between blocks and their siblings.

As part of this, we now try to apply spacing on all `ParseElement`
invocations. Previuosly, we would only do this in specific cases, but
now that this has been made more consistent, we can apply it more
generally and expect the correct result.