EES-4577 Add fallback validation and submit errors and show by default
This change re-works our forms so that we can configure error messages more
precisely for the following cases:

- When no server validation errors mapped to form fields
- When non-server validation errors thrown during form submit

We can now control the messages for these using fallback messages, or if we
want a more precise message, we can throw a `SubmitError` with a specific
message (and even a field).

As part of this, we no longer need the `showSubmitError` option as we can always
display a user-friendly error message. This is more useful than the previous
behaviour that could potentially show unfriendly error messages.