EES-4611 Re-implement request count tracking using state setter callbacks
This change re-implements request count tracking so that we use the state setter
callbacks instead of a combination of refs and state setters.

Hopefully, this should provide a bit more atomicity, whereas previously it would
be possible for the request count to change before it had been committed to state.