EES-4798 Add customised problem details for built-in data annotations
This allows customised validation problem details to be used for the
older projects that are still using the data annotations instead of
FluentValidation. However, this isn't ideal as validation errors from
data annotations can't have codes or extra details.

For now, we just include the path and message, which provides a bit of
consistency in the error response for all APIs.