EES-4860 Normalise data set meta into tables (instead of JSONB)
This changes replaces the denormalised JSONB we had initially
implemented for the various data set meta with standard SQL tables.

Unfortunately, the JSONB approach had poor performance characteristics
when trying to SELECT options. In these cases, we have to convert the
JSONB array into table rows so they can be returned by the
query. This is imprecise and means Postgres has to potentially read and
transform megabytes of JSONB (e.g. schools, qualifications, etc).

The normalised table approach is much more performant as the meta is
actually very tabular in nature. This means we can have efficient
indices on the columns, making SELECTs very fast.

Additionally, we have implemented many-to-many link tables with
immutable option tables that should significantly reduce the storage
overhead. This will most likely save space compared to JSONB as we won't
store nearly as much duplicated / redundant data.

As per the main objective of this work, we've also updated the location
meta to more correctly model how locations actually work. This means
that there are specialised types of locations that have one or more
special codes that can be used. We'll eventually allow users to search
on all these columns.