EES-4716 Allow `SeparatedQueryModelBinder` to bind on string lists via attribute
This changes `SeparatedQueryModelBinder` so that it can bind enumerable
string types from a query parameter containing comma separated strings
e.g. `?items=option-1,option-2`.

Previously, this was not possible as the model binder would ignore these
model types on the basis that strings often contain commas without being
intended to be used as comma-separated list.

We can now enable binding to enumerable string types by adding the
`QuerySeparator` attribute to the property or parameter. Note that
`FromQuery` is also required.