EES-5017 Fix meta summary's `GeographicLevels` not being populated correctly
This changes the `GeographicLevels` to be correctly populated from the
`GeographicLevelMeta` instead of the `LocationMetas`.

As part of this, we've also updated `DataSetVersion` generator
extensions to set the geographic levels meta based on the locations meta
in the default cases. This is sensible default to avoid having to
explicitly set the geographic levels every time.