EES-5015 Refactor `ParquetPathResolver` and associated utils and paths
This refactors `ParquetPathResolver` to use default interface
implementations. This should make it easier to create implementations
for test projects where we need more control over the paths.

As part of this, we've essentially removed all the Parquet path utils and
extensions as we should force `IParquetPathResolver` to be used instead.
Filename constants have been refactored into the respective table classes
to reflect this.