EES-5057 Add `ProblemDetails` body for 403 Forbidden responses
Previously, .NET would not output a response body for a 403 Forbidden
as they leave it up to you do what you want with such a response (e.g.
redirect the browser to a login page).

However, this isn't helpful in the context of a REST API, so we've now
fixed this to return a default `ProblemDetails` response.