EES-5074 Fix streaming endpoints not returning ActionResults
This fix is implemented using a `NoOpResult` which allows us to process
the streaming request in the controller without any further modification
by the framework.

Previously, returning any other type of `ActionResult` would result in
modifications to the response. This would cause exceptions due to the
stream modifying the response after it has already started.