EES-4370 Wrapping the trigger in a try/catch to prevent hanging when an exception is thrown
This is a temporary fix. The cause of this bug is unknown. The bug causes the request to hang until a timeout occurs when an exception is thrown. There is an issue here (https://github.com/Azure/azure-functions-dotnet-worker/issues/2476) which we are tracking. If the problem gets fixed, then we will unwrap the trigger.