EES-4906 Add default `DataFixture` instances for `Generators`
Anytime a `Generator` is instantiated, if it is not provided a
`DataFixture` instance, it will set up its own. This has a major benefit
of always have an in-context `DataFixture` meaning that the various
fixture indices are always populated (instead of being potentially -1).

This also allows us pass the `DataFixture` down into any sub-contexts
e.g. when setting child relationships, and have consistent fixture indices.