EES-5173 Fixing bug whereby we already deleted the `ReleaseFiles` and `Files` before attempting to delete the API data sets attached to those + Adding a check to block deleting if the plan is not valid
Prior to these changes, we could, under certain circumstances, proceed with deleting a release version even if the deletion plan was not valid. There was also a bug whereby we would delete all the `Files` and `ReleaseFiles` for the `ReleaseVersion` first, and then make a call to the Processor to delete the API Data Sets. When this happened, the Processor could not find the `DataSetVersions` linked to the `ReleaseVersion` to delete them, because the `ReleaseFiles` had already been deleted.