EES-5173 Removing changes to the deletion plan
We discussed and decided that we do not need to check which API data sets are linked to the release version when deleting it. Prior to these changes, we just allowed the deletion to go through without checking the plan. The plan did not validate anything. So, we've made the decision to simplify it and continue to not check against the plan when deleting the release version.