EES-5173 Change Release Version DELETION plan to include deleting draft API data sets (#4957)
* EES-5173 Changing the Release Version DELETION plan to include deleting draft API data sets

* EES-5173 Changing the DELETE release endpoint to not convert `NotFoundResult` into `NoContentResult`

* EES-5173 Adding tests

* EES-5173 Fixing bug whereby we already deleted the `ReleaseFiles` and `Files` before attempting to delete the API data sets attached to those + Adding a check to block deleting if the plan is not valid

Prior to these changes, we could, under certain circumstances, proceed with deleting a release version even if the deletion plan was not valid. There was also a bug whereby we would delete all the `Files` and `ReleaseFiles` for the `ReleaseVersion` first, and then make a call to the Processor to delete the API Data Sets. When this happened, the Processor could not find the `DataSetVersions` linked to the `ReleaseVersion` to delete them, because the `ReleaseFiles` had already been deleted.

* EES-5173 Addressing PR comments

* EES-5173 Deserialising `Status` enum as a string

* EES-5173 Removing changes to the deletion plan

We discussed and decided that we do not need to check which API data sets are linked to the release version when deleting it. Prior to these changes, we just allowed the deletion to go through without checking the plan. The plan did not validate anything. So, we've made the decision to simplify it and continue to not check against the plan when deleting the release version.

* EES-5173 Some cleanup

* EES-5173 Fixing tests

* EES-5173 Removing unused method