EES-4950 Re-model changelog data model for public API
This re-models the changelog data models by:

- Removing `ChangeSets`
- Removing JSONB columns in favour of a standard relational model using
  foreign keys (i.e. previous / current states).

We've moved towards this approach as it allows us to avoid duplicating
the entire meta domain in JSONB columns (and respective models).
Using normal foreign key references means we can keep things relatively
simple.

The only issue is that we don't know if this approach will perform well
at scale, but we believe it'd be possibel to remedy this with blob
caching (or similar).