EES-4928 Adding tests + Refactoring the `EmailService`
The `INotificationClientProvider` was somewhat unnecessary, and caused the need for unnecessary setup in both the business logic and tests. We have simplified this by removing this class and injecting the `INotificiationClient` at startup. Only the `EmailClient` needs access to this - so now we only need to mock the `IEmailService` in our integration tests (because we don't want to run our tests against the real Notify project, which is a third-party dependency).